Create postcss-import@8.0.2.json
See here: https://github.com/postcss/postcss-import/issues/181

`load-content` errors out when attempting to run in the browser. This isn't a big deal as it's module is overridden with options passed into PostCSS-Import. However, the module needs to not error out. So, I add an override for the browser environment which ensures an empty response.

QUESTION:

It's not clear to me whether I need to specify all other mapped dependencies in this file, or if I should only specify the one property.

Also, it's not clear to me how I would go about submitting a PR for something like this to postcss-import. I know this registry is supposed to be a temp location, but this seems specific to JSPM and not to their plugin.