feat(@angular/*): use different main entry in production mode
When setting production to true, the main entry is set to the index.js file.
The file (like the other files in the src directory) is an ES5 JavaScript file
with ES6 module format. Using this file instead of the UMD bundle allows
for tree-shaking when building the package in production mode using
rollup.